requirements.txt
streamlit
easyocr
Pillow
torch
opencv-python-headless
imageio[pyav]
imageio[ffmpeg]
datasets
pytesseract
matplotlib
language-tool-python
textblob
spellchecker